requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
pytest==7.0.1
playwright
pytest-xdist==1.32.0
pytest-rerunfailures==10.2
retry==0.9.2

